+++ src/Rhetos.Core.Integration/Deployment/InstalledPackage.cs
@@ -17,11 +17,9 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using System.IO;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Text;
-using System;
+using System.IO;
 using System.Linq;
 
 namespace Rhetos.Deployment
@@ -36,22 +34,22 @@ namespace Rhetos.Deployment
         public InstalledPackage(
             string id,
             string version,
-            IEnumerable<PackageRequest> dependencies,
+            List<PackageRequest> dependencies,
             string folder,
             List<ContentFile> contentFiles)
         {
             Id = id;
             Version = version;
-            Dependencies = dependencies;
+            Dependencies = dependencies ?? new();
             Folder = folder;
-            ContentFiles = contentFiles;
+            ContentFiles = contentFiles ?? new();
         }
 
         public string Id { get; private set; }
 
         public string Version { get; private set; }
 
-        public IEnumerable<PackageRequest> Dependencies { get; private set; }
+        public List<PackageRequest> Dependencies { get; private set; }
 
         /// <summary>
         /// The local cache folder at build-time, where the package files are extracted and used by Rhetos.
@@ -77,56 +75,6 @@ namespace Rhetos.Deployment
                 file.InPackagePath = file.InPackagePath
                     .Replace('/', Path.DirectorySeparatorChar)
                     .Replace('\\', Path.DirectorySeparatorChar);
-        }
-
-        const string DslScriptsSubfolder = "DslScripts";
-        private static readonly string DslScriptsSubfolderPrefix = DslScriptsSubfolder + Path.DirectorySeparatorChar;
-
-        /// <summary>
-        /// Extracts all files from a given subfolder into a new (virtual) package, and removes them from the current package.
-        /// </summary>
-        internal InstalledPackage ExtractSubpackage(Subpackage subpackage)
-        {
-            string subpackageFolder = Path.GetFullPath(Path.Combine(Folder, subpackage.Folder));
-            if (subpackageFolder.Last() != Path.DirectorySeparatorChar)
-                subpackageFolder += Path.DirectorySeparatorChar; // Makes sure to avoid selecting files from a subfolder which Name begins with the wanted subfolder name.
-
-            subpackage.Dependencies ??= Array.Empty<string>();
-
-            var virtualPackage = new InstalledPackage(
-                Id + "." + subpackage.Name,
-                "",
-                Dependencies.Concat(subpackage.Dependencies.Select(dependency => new PackageRequest { Id = Id + "." + dependency, VersionsRange = "" })).ToList(),
-                subpackageFolder,
-                ContentFiles.Where(f => f.PhysicalPath.StartsWith(subpackageFolder, StringComparison.OrdinalIgnoreCase)).ToList());
-
-            ContentFiles.RemoveAll(f => f.PhysicalPath.StartsWith(subpackageFolder, StringComparison.OrdinalIgnoreCase));
-
-            foreach (var file in virtualPackage.ContentFiles)
-            {
-                // Files in the new packages should have paths relative to the package's folder, instead of the project's root folder.
-                if (file.InPackagePath.Length <= subpackage.Folder.Length
-                    || (file.InPackagePath[subpackage.Folder.Length] != Path.DirectorySeparatorChar
-                    && file.InPackagePath[subpackage.Folder.Length] != Path.AltDirectorySeparatorChar))
-                    throw new FrameworkException($"Unexpected InPackagePath of a file '{file.InPackagePath}'." +
-                        $" It should start with the subpackage Folder name '{subpackage.Folder}' followed by a directory separator.");
-                file.InPackagePath = file.InPackagePath.Substring(subpackage.Folder.Length + 1);
-
-                // DSL scripts in the referenced packages need to be in the DslScripts folder, to match the behavior of DiskDslScriptLoader.LoadPackageScripts.
-                if (Path.GetExtension(file.InPackagePath).Equals(".rhe", StringComparison.OrdinalIgnoreCase))
-                    if (!file.InPackagePath.StartsWith(DslScriptsSubfolderPrefix, StringComparison.OrdinalIgnoreCase))
-                        file.InPackagePath = Path.Combine(DslScriptsSubfolder, file.InPackagePath);
-            }
-
-            if (!virtualPackage.ContentFiles.Any() && !Directory.Exists(subpackageFolder))
-                throw new ArgumentException($"Subpackage '{subpackage.Name}' directory '{subpackageFolder}' does not exist. Review the Rhetos build settings.");
-
-            return virtualPackage;
-        }
-
-        internal void AddDependencies(List<InstalledPackage> createdPackages)
-        {
-            Dependencies = Dependencies.Concat(createdPackages.Select(p => new PackageRequest { Id = p.Id, VersionsRange = "" })).ToList();
         }
     }
 }