+++ test/TestAppDirectMsSqlEfCore/Program.cs
@@ -29,22 +29,23 @@ IHostBuilder builder = Host.CreateDefaultBuilder(args);
 builder.ConfigureServices(services =>
 {
     services.AddRhetosHost((IServiceProvider serviceProvider, IRhetosHostBuilder rhetosHostBuilder) =>
-        {
-            rhetosHostBuilder
-                .ConfigureRhetosAppDefaults()
-                .ConfigureConfiguration(configurationBuilder => configurationBuilder
-                    .AddJsonFile("local.settings.json"))
-                .ConfigureContainer(builder =>
-                {
-                    builder.RegisterType<ProcessUserInfo>().As<IUserInfo>();
-                });
-        });
+    {
+        rhetosHostBuilder
+            .ConfigureRhetosAppDefaults()
+            .ConfigureConfiguration(configurationBuilder => configurationBuilder
+                .AddJsonFile("local.settings.json"))
+            .ConfigureContainer(builder =>
+            {
+                builder.RegisterType<ProcessUserInfo>().As<IUserInfo>();
+            });
+    });
 });
 
 IHost host = builder.Build();
 
 var rhetosAppOptions = host.Services.GetService<IRhetosComponent<RhetosAppOptions>>()?.Value;
 Console.WriteLine($"Running Rhetos app '{rhetosAppOptions?.RhetosAppAssemblyFileName}'.");
+Console.WriteLine($"Generated classes in assembly '{typeof(Bookstore.Book).Assembly.GetName().Name}'.");
 
 Console.WriteLine("Executing test query:");
 var repository = host.Services.GetService<IRhetosComponent<Common.DomRepository>>().Value;